+++ example/src/App.js
@@ -5,17 +5,23 @@ import { DatePicker } from 'simplifa11y'
 const App = () => {
   const [dateString, setDateString] = useState('')
 
+  // const theme = {
+  //   dateInput: {
+  //     wrap: {
+  //       width: '250px'
+  //     },
+  //     input: {
+  //       padding: '0 0 0 20px'
+  //     },
+  //     inputHint: {
+  //       padding: '0 0 0 20px'
+  //     }
+  //   }
+  // }
+
   const theme = {
-    dateInput: {
-      wrap: {
-        width: '250px'
-      },
-      input: {
-        padding: '0 0 0 20px'
-      },
-      inputHint: {
-        padding: '0 0 0 20px'
-      }
+    calendar: {
+      zIndex: '2'
     }
   }
 
@@ -28,8 +34,8 @@ const App = () => {
         formatHint='MM DD YYYY'
         onSelect={setDateString}
         dateRange={{
-          min: '08/02/2021',
-          max: '09/20/2021'
+          min: '08/02/2002',
+          max: '12/31/2008'
         }}
       />